requirements.txt
pdfplumber
python-pptx
moviepy
gTTS
Pillow